Separacion de componentes Admin y Customer Panel
