+++ Frontend/art-gallery/src/components/GeneralPanel/GeneralPanel.jsx
@@ -1,7 +1,8 @@
-import React from 'react'
+import React from "react";
 import HeaderCheckPanel from "../utils/HeaderCheckPanel/HeaderPanel";
 import SubtitlePanel from "../utils/SubtitlePanel/SubtitlePanel";
-import CardPanelGeneral from '../utils/CardPanelGeneral/CardPanelGeneral';
+import CardPanelCustomer from "../utils/CardPanelGeneral/CardPanelCustomer";
+import CardPanelAdmin from "../utils/CardPanelGeneral/CardPanelAdmin";
 import useAuth from "../../customHooks/useAuth";
 import "./GeneralPanel.css";
 function GeneralPanel() {
@@ -11,11 +12,11 @@ function GeneralPanel() {
   const title = user.role === "admin" ? "Sales" : "Purchase";
   return (
     <div className="sales-container">
-        <HeaderCheckPanel title={title} />
-        <SubtitlePanel isAdmin={isAdmin} />
-        <CardPanelGeneral isAdmin={isAdmin}/>
-      </div>
-  )
+      <HeaderCheckPanel title={title} />
+      <SubtitlePanel isAdmin={isAdmin} />
+      {isAdmin === true ? <CardPanelAdmin /> : <CardPanelCustomer />}
+    </div>
+  );
 }
 
-export default GeneralPanel
+export default GeneralPanel;
